// src/app/Medical/Models/LTCSignaturemap.ts
export class LTCSignatureMapData {
    ID: any;
    LTC_ID: any;
    LTC_ORDER_SIGNATURE_1: any;
    LTC_ORDER_SIGNATURE_2: any;
    LTC_ORDER_SIGNATURE_3: any;
    LTC_PART_B_SIGNATURE_1: any;
    LTC_PART_B_SIGNATURE_2: any;
    LTC_PART_B_SIGNATURE_3: any;
    LTC_PART_B_SIGNATURE_4: any;
    LTC_PART_B_SIGNATURE_5: any;
    ADV_LTC_GAR_SIGNATURE_1: any;
    ADV_LTC_GAR_SIGNATURE_2: any;
    ADV_LTC_MANJURI_ADESH_SIGNATURE_1: any;
    ADV_LTC_MANJURI_ADESH_SIGNATURE_2: any;
    ADV_LTC_ADV_ORDER_SIGNATURE_1: any;
    ADV_LTC_ADV_ORDER_SIGNATURE_2: any;
}